Convert all of the Markdown to HTML
On Reddit Kann-Ich was asking about the differences between heading. I
decided to start supporting more Markdown and ended up using
showdown[0]. This is a better approach then dicking around with Regex to
get the various mappings right. Now headings and nested markdown should
work. It even works on the deck name :)

I also discovered another bug where image matching could break, so found
a different RegeExp for that. I will likely rewrite the mangling of the
media urls later. Did not want to deal with HTML parsing for now and the
replace for markdown is simple currently.

Also set the allowed filetype to zip for now.

[0]: https://github.com/showdownjs/showdown