fix blank page on firefox
The fix is just to use different regular expression for images for now.
I don't know what the differences between Chrome and Firefox are but
there are some.

I really should clean this up but kind of limited on time so this will
have to do for now. Will revisit this in the next week, so hopefully it
does not break any use case.

Closes:  #2 ([BUG] Page doesn’t render under Firefox)
Reported-by: andrewcrook