parser: wrap the initial proof of concept for tag support
This is assuming that the user is setting it via the strikethrough which
seems to be a good idea, but who knows. We will likely get some feedback
very soon.

Here is my notes from the experimentation:
> Here is my cool idea for adding tags support on the cards. We can use
the strike through which is very distinct and easy to communicate. Not
only that it even has a shortcut! On windows that is `Ctrl+Shift+S` on
macOS that is `Cmd+Shift+S`.