genanki: parser: views: expose the template names to the user
We want to give the user the full powers. They should be able to
generate new templates. This is possible by changing the names so you
can avoid affecting your previous imports / flashcards.

Thanks to Abhiyan Bhandari for showing is use case on [How to Make Your
Anki Cards Look Beautiful (+FREE STYLING TEMPLATE)][0] that convinced me
it's time to get this out. We will revisit the interface at a later
stage when the preview support comes.

[0]: https://youtu.be/D0GFO6NnMLM