web: save the last selected filename to local storage
Note that we still need to resolve the issue where the input is being
removed from the DOM. From what I read on Stack Overflow[0] it's not
possible to programmatically set the input file. The user has to select
it.

Will try something else where the DOM is unchanged and we show something
on top like a modal.

[0]: https://stackoverflow.com/questions/51422294/how-to-hold-file-input-value-in-react