fix: bookmarks not showing up when request fails
We are simplifying the implementation to no longer guess the description
based on the contents of the page. Doing that does work when the page
has paragraphs but for customized layouts there are no grantees.
Defaulting to showing title and URL when no meta tags are present is
consistent with how Notion does it.

Closes: https://github.com/alemayhu/2anki.net/issues/555
\#555: Bookmarks not working when using multiple bookmarks