fix: resolve the issue with invalid zip data
When the zip file can not be unzipped we attempt to use `tar` as
fallback mechanism. After testing this for a while, can confirm the
fallback is good enough but it cannot replace the zip handler altogether
like attempted before in:

- https://github.com/2anki/server/pull/1487
- https://github.com/2anki/server/pull/1495

Instead this approach will only use it if we receive the invalid zip
exception from fflate.